requirements.txt
altgraph==0.17.3
blinker==1.6.2
bottle==0.12.25
click==8.1.3
Flask==2.3.2
Flask-Cors==3.0.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
proxy-tools==0.1.0
pycairo==1.24.0
pyinstaller==5.12.0
pyinstaller-hooks-contrib==2023.3
pylogix==0.8.14
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
pywebview==4.1
QtPy==2.3.1
six==1.16.0
Werkzeug==2.3.6
